# frontend/src/generated/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_resume_api_v1_resumes_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_v1_resumes_upload_post",
        "type": "object"
      },
...
      "Contact": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "links": {
            "items": {
              "type": "string"
            },
            "title": "Links",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name"
        ],
        "title": "Contact",
        "type": "object"
      },
...
      "Country": {
        "enum": [
          "US",
          "IN"
        ],
        "title": "Country",
        "type": "string"
      },
...
      "EvalRequest": {
        "properties": {
          "target_role": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Target Role",
            "type": "string"
          }
        },
        "required": [
          "target_role"
        ],
        "title": "EvalRequest",
        "type": "object"
      },
...
      "ExportRequest": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Country"
              },
              {
                "type": "null"
              }
            ]
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "resume_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Document Id"
          },
          "resume_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Version Id"
          },
          "role_template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoleTemplate"
              },
              {
                "type": "null"
              }
            ]
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "title": "ExportRequest",
        "type": "object"
      },
      "ExportResponse": {
        "properties": {
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "export_id": {
            "title": "Export Id",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "role_template": {
            "$ref": "#/components/schemas/RoleTemplate"
          }
        },
        "required": [
          "export_id",
          "download_url",
          "expires_at",
          "country",
          "role_template"
        ],
        "title": "ExportResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumeDetailResponse": {
        "properties": {
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumeSavedEvaluation"
              },
              {
                "type": "null"
              }
            ]
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeListItem"
          }
        },
        "required": [
          "resume",
          "evaluation"
        ],
        "title": "ResumeDetailResponse",
        "type": "object"
      },
      "ResumeEvaluationResponse": {
        "properties": {
          "ats_parseability": {
            "title": "Ats Parseability",
            "type": "integer"
          },
          "ats_raw_text": {
            "title": "Ats Raw Text",
            "type": "string"
          },
          "bullet_flags": {
            "items": {
              "$ref": "#/components/schemas/BulletFlag"
            },
            "title": "Bullet Flags",
            "type": "array"
          },
          "evaluation_id": {
            "title": "Evaluation Id",
            "type": "string"
          },
          "format_issues": {
            "items": {
              "$ref": "#/components/schemas/FormatIssue"
            },
            "title": "Format Issues",
            "type": "array"
          },
          "overall_score": {
            "maximum": 100,
            "minimum": 0,
            "title": "Overall Score",
            "type": "integer"
          },
          "parser_confidence": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Parser Confidence",
            "type": "string"
          },
          "readiness_label": {
            "enum": [
              "ready",
              "minor_edits",
              "needs_work"
            ],
            "title": "Readiness Label",
            "type": "string"
          },
          "score_breakdown": {
            "$ref": "#/components/schemas/ScoreBreakdown"
          },
          "score_explanation": {
            "items": {
              "$ref": "#/components/schemas/ScoreExplanation"
            },
            "title": "Score Explanation",
            "type": "array"
          },
          "summary_critique": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Critique"
          },
          "top_actions_before_applying": {
            "items": {
              "type": "string"
            },
            "title": "Top Actions Before Applying",
            "type": "array"
          }
        },
        "required": [
          "overall_score",
          "readiness_label",
          "score_breakdown",
          "score_explanation",
          "top_actions_before_applying",
          "parser_confidence",
          "bullet_flags",
          "format_issues",
          "evaluation_id",
          "summary_critique",
          "ats_parseability",
          "ats_raw_text"
        ],
        "title": "ResumeEvaluationResponse",
        "type": "object"
      },
...
      "ResumeListItem": {
        "properties": {
          "evaluation_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumeSavedEvaluation"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluation_status": {
            "enum": [
              "pending",
              "completed"
            ],
            "title": "Evaluation Status",
            "type": "string"
          },
          "file_size": {
            "title": "File Size",
            "type": "integer"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_primary": {
            "title": "Is Primary",
            "type": "boolean"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "resume_document_id": {
            "title": "Resume Document Id",
            "type": "string"
          },
          "uploaded_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded At"
          }
        },
        "required": [
          "id",
          "resume_document_id",
          "filename",
          "original_filename",
          "file_size",
          "file_type",
          "uploaded_at",
          "evaluation_status",
          "is_primary",
          "evaluation_result"
        ],
        "title": "ResumeListItem",
        "type": "object"
      },
      "ResumeListResponse": {
        "properties": {
          "resumes": {
            "items": {
              "$ref": "#/components/schemas/ResumeListItem"
            },
            "title": "Resumes",
            "type": "array"
          },
          "totalCount": {
            "title": "Totalcount",
            "type": "integer"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "resumes",
          "total_count",
          "totalCount"
        ],
        "title": "ResumeListResponse",
        "type": "object"
      },
...
      "ResumeUploadResponse": {
        "properties": {
          "certifications": {
            "items": {
              "type": "string"
            },
            "title": "Certifications",
            "type": "array"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationEntry"
            },
            "title": "Education",
            "type": "array"
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/ExperienceEntry"
            },
            "title": "Experience",
            "type": "array"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectEntry"
            },
            "title": "Projects",
            "type": "array"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          },
          "resume_document_id": {
            "title": "Resume Document Id",
            "type": "string"
          },
          "skills": {
            "$ref": "#/components/schemas/Skills"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "contact",
          "raw_text",
          "resume_document_id"
        ],
        "title": "ResumeUploadResponse",
        "type": "object"
      },
      "ResumeVersionResponse": {
        "properties": {
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "version_id"
        ],
        "title": "ResumeVersionResponse",
        "type": "object"
      },
      "RewriteRequest": {
        "properties": {
          "country": {
            "default": "US",
            "title": "Country",
            "type": "string"
          },
          "jd_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jd Context"
          },
          "target_role": {
            "title": "Target Role",
            "type": "string"
          }
        },
        "required": [
          "target_role"
        ],
        "title": "RewriteRequest",
        "type": "object"
      },
      "RewriteResult": {
        "properties": {
          "applied_changes": {
            "items": {
              "type": "string"
            },
            "title": "Applied Changes",
            "type": "array"
          },
          "placeholders": {
            "items": {
              "$ref": "#/components/schemas/Placeholder"
            },
            "title": "Placeholders",
            "type": "array"
          },
          "rewritten": {
            "title": "Rewritten",
            "type": "string"
          }
        },
        "required": [
          "rewritten",
          "placeholders",
          "applied_changes"
        ],
        "title": "RewriteResult",
        "type": "object"
      },
      "RoleTemplate": {
        "enum": [
          "swe",
          "ds",
          "pm"
        ],
        "title": "RoleTemplate",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionRequest": {
        "properties": {
          "change_set": {
            "items": {
              "$ref": "#/components/schemas/ChangeItem"
            },
            "title": "Change Set",
            "type": "array"
          },
          "parent_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version Id"
          }
        },
        "required": [
          "change_set"
        ],
        "title": "VersionRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-Powered Job Extraction & Management API",
    "title": "Prism Pro - Job Extraction & Management",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/exports": {
      "post": {
        "operationId": "create_export_api_v1_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Export",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/{export_id}": {
      "get": {
        "operationId": "get_export_api_v1_exports__export_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "title": "Export Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Export",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/{export_id}/download": {
      "get": {
        "operationId": "download_export_api_v1_exports__export_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "title": "Export Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "private, no-store",
                "schema": {
                  "type": "string"
                }
              },
              "Content-Disposition": {
                "description": "PDF attachment filename",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Export",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/api/v1/resumes/list": {
      "get": {
        "operationId": "list_resumes_api_v1_resumes_list_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Resumes",
        "tags": [
          "resumes-v2",
          "resumes-v2"
        ]
      }
    },
    "/api/v1/resumes/upload": {
      "post": {
        "operationId": "upload_resume_api_v1_resumes_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_v1_resumes_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Resume",
        "tags": [
          "resumes-v2",
          "resumes-v2"
        ]
      }
    },
    "/api/v1/resumes/{resume_document_id}": {
      "delete": {
        "operationId": "delete_resume_api_v1_resumes__resume_document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "resume_document_id",
            "required": true,
            "schema": {
              "title": "Resume Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Resume",
        "tags": [
          "resumes-v2",
          "resumes-v2"
        ]
      },
      "get": {
        "operationId": "get_resume_api_v1_resumes__resume_document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "resume_document_id",
            "required": true,
            "schema": {
              "title": "Resume Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Resume",
        "tags": [
          "resumes-v2",
          "resumes-v2"
        ]
      }
    },
    "/api/v1/resumes/{resume_document_id}/evaluate": {
      "post": {
        "operationId": "evaluate_api_v1_resumes__resume_document_id__evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "resume_document_id",
            "required": true,
            "schema": {
              "title": "Resume Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeEvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evaluate",
        "tags": [
          "resumes-v2",
          "resumes-v2"
        ]
      }
    },
    "/api/v1/resumes/{resume_document_id}/rewrite/{bullet_id}": {
      "post": {
        "operationId": "rewrite_api_v1_resumes__resume_document_id__rewrite__bullet_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "resume_document_id",
            "required": true,
            "schema": {
              "title": "Resume Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bullet_id",
            "required": true,
            "schema": {
              "title": "Bullet Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewriteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rewrite",
        "tags": [
          "resumes-v2",
          "resumes-v2"
        ]
      }
    },
    "/api/v1/resumes/{resume_document_id}/versions": {
      "post": {
        "operationId": "create_version_api_v1_resumes__resume_document_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "resume_document_id",
            "required": true,
            "schema": {
              "title": "Resume Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Version",
        "tags": [
          "resumes-v2",
          "resumes-v2"
        ]
      }
    },
...
  }
}